Fix time zone bug
The Discord bot would send the Word of the Day at 8:00am Pacific Time, which has daylight savings. The script to update the countdown didn't set the time zone with pytz, but instead got the time from UTC. During daylight savings, the bot's activity status would be offset an hour from when the Word of the Day is actually sent.